Move user model and service into the users module, scroll service into the shared module; remove unused displacer component/directive
