+++ app/authentication/login/login.component.ts
@@ -2,7 +2,7 @@ import { Component } from '@angular/core';
 import { Response } from '@angular/http';
 
 import { AuthService } from '../../core/auth/auth.service';
-import { User } from '../../core/user/user.model';
+import { User } from '../../users/user.model';
 
 @Component({
   moduleId: module.id,
