+++ app/authentication/signup/signup.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 
 import { AuthService } from '../../core/auth/auth.service';
-import { User } from '../../core/user/user.model';
+import { User } from '../../users/user.model';
 
 @Component({
   moduleId: module.id,
