+++ app/core/guards/admin-guard.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { ActivatedRouteSnapshot, CanActivate, Router, RouterStateSnapshot } from '@angular/router';
 
-import { UserService } from '../user/user.service';
+import { UserService } from '../../users/user.service';
 
 @Injectable()
 export class AdminGuard implements CanActivate {
