+++ app/shared/taginput/user-tag-input.component.ts
@@ -3,8 +3,8 @@
 } from '@angular/core';
 import { Observable } from 'rxjs/Rx';
 
-import { User } from '../../core/user/user.model';
-import { UserService } from '../../core/user/user.service';
+import { User } from '../../users/user.model';
+import { UserService } from '../../users/user.service';
 
 @Component({
   moduleId: module.id,
