+++ app/topics/show-topic/show-topic.component.ts
@@ -6,8 +6,8 @@ import { Subscription } from 'rxjs';
 
 import { Topic } from '../shared/topic.model';
 import { TopicService } from '../shared/topic.service';
-import { User } from '../../core/user/user.model';
-import { UserService } from '../../core/user/user.service';
+import { User } from '../../users/user.model';
+import { UserService } from '../../users/user.service';
 
 @Component({
   moduleId: module.id,
