+++ app/users/admin/pipes/sort.pipe.ts
@@ -1,6 +1,6 @@
 ﻿import { Injectable, Pipe, PipeTransform } from '@angular/core';
 
-import { User } from '../../../core/user/user.model';
+import { User } from '../../user.model';
 
 @Pipe({
   name: 'hipUsersSorter'
