+++ app/users/admin/users-list/users-list.component.ts
@@ -1,7 +1,7 @@
 ﻿import { Component, OnInit } from '@angular/core';
 
-import { UserService } from '../../../core/user/user.service';
-import { User } from '../../../core/user/user.model';
+import { UserService } from '../../user.service';
+import { User } from '../../user.model';
 import { Roles } from '../roles.model';
 
 @Component({
